+++ src/config/site.config.ts
@@ -1,0 +1,21 @@
+export const siteConfig = {
+	author: {
+		name: 'Nicolás Deyros',
+		roles: [
+			'Developer',
+			'Project Manager',
+			'Advertiser',
+			'a.k.a Problem Solver',
+		],
+		email: 'contact@ndeyros.dev', // Fallback/Public email
+	},
+	email: {
+		subjectPrefix: 'Contact Form:',
+		senderName: 'Portfolio Contact Form',
+	},
+	meta: {
+		title: 'Nicolás Deyros - Portfolio',
+		description:
+			'Welcome to my personal website. Here you can find information about me, my projects, and how to contact me.',
+	},
+}